[loop_tool] Integrate loop tool scaffolding into build system.
