+++ compiler_gym/envs/loop_tool/service/BUILD
@@ -1,0 +1,13 @@
+# Copyright (c) Facebook, Inc. and its affiliates.
+#
+# This source code is licensed under the MIT license found in the
+# LICENSE file in the root directory of this source tree.
+
+filegroup(
+    name = "service",
+    srcs = [
+        "compiler_gym-loop_tool-service",
+        "loop_tool_compilation_session.py",
+    ],
+    visibility = ["//compiler_gym/envs/loop_tool:__subpackages__"],
+)
