+++ compiler_gym/envs/loop_tool/service/compiler_gym-loop_tool-service
@@ -1,0 +1,13 @@
+#! /usr/bin/env python3
+#
+#  Copyright (c) Facebook, Inc. and its affiliates.
+#
+# This source code is licensed under the MIT license found in the
+# LICENSE file in the root directory of this source tree.
+from compiler_gym.envs.loop_tool.service.loop_tool_compilation_session import (
+    LoopToolCompilationSession,
+)
+from compiler_gym.service.runtime import create_and_run_compiler_gym_service
+
+if __name__ == "__main__":
+    create_and_run_compiler_gym_service(LoopToolCompilationSession)
